Use given local addr for outgoing response, closes #5
When a server has multiple IP, disregaring which IP was used to make the
connexion, the response will use the wildcard IP as source.

It cause some issues with certain PXE firmaware.

Example :

The TFTP server has IP 192.168.1.1 and 192.168.1.2. The primary IP is .1 but we
want to use .2 for the TFTP server.

The cLient has IP 192.168.1.3

The machine sends a request 192.168.1.3 => 192.168.1.2 and the response is from
192.168.1.1. Several servers reject this.

With this patch, if NewTFTPServer is called with a specific IP address, this IP
will be used as source for the responses. If the wildcard is used, nothing
will change.

NOTE: patch originally created by Xavier Henner (https://github.com/xhenner) on
PR #5.